IDEA-160427 IDEA-CR-13371 now cast to (long) is added (probably sometimes excessive) in order to produce correct code

+++ java/java-tests/testData/inspection/streamApiCallChains/afterStreamCount.java
@@ -4,6 +4,6 @@ import java.util.Arrays;
 
 class Test {
   long cnt() {
-    return Arrays.asList('d', 'e', 'f').size();
+    return (long) Arrays.asList('d', 'e', 'f').size();
   }
 }
